MLBackend/requirements.txt
flask==2.3.3
flask-socketio==5.3.6
tensorflow==2.18.0
scipy==1.11.1
pyserial==3.5
python-socketio==5.8.0
eventlet==0.33.3 